python/requirements.txt
# Python dependencies for SweetDesk
# Install with: pip install -r requirements.txt

numpy>=1.24.0
Pillow>=10.0.0

# Optional: For production-grade implementations
# imgutils>=0.5.0  # DeepGHS image classification
# seam-carving>=1.1.0  # Better seam carving implementation
